requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.8.1
assemblyai==0.37.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioread==3.0.1
autobahn==24.4.2
Automat==24.8.1
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels-redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
daphne==4.1.2
dataclasses-json==0.6.7
decorator==5.2.1
deepgram-sdk==2.12.0
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
docx2txt==0.8
einops==0.8.1
faiss-cpu==1.10.0
ffmpeg-python==0.2.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gevent==24.11.1
google-auth==2.38.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
hyperlink==21.0.0
HyperPyYAML==1.2.2
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
julius==0.2.7
kiwisolver==1.4.8
langchain==0.3.21
langchain-core==0.3.48
langchain-text-splitters==0.3.7
langcodes==3.5.0
langsmith==0.3.18
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.1
lightning-utilities==0.14.2
llvmlite==0.44.0
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.69.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
optuna==4.2.1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
preshed==3.0.9
primePy==1.3
propcache==0.3.0
protobuf==6.30.1
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.0
pydantic_core==2.33.0
pyfcm==2.0.7
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyPDF2==3.0.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
service-identity==24.2.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.12.1
sox==1.5.0
soxr==0.5.0.post1
spacy==3.8.4
speechbrain==1.0.2
SpeechRecognition==3.14.2
SQLAlchemy==2.0.39
sqlparse==0.5.3
srsly==2.5.1
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.7.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
websocket-client==1.8.0
websockets>=11.0
whisper==1.1.10
zstandard==0.23.0
